requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6